+++ heat/core/tests/test_manipulations.py
@@ -1307,34 +1307,30 @@ class TestManipulations(BasicTest):
         self.assertTrue((result._DNDarray__array == data._DNDarray__array.squeeze()).all())
 
         # 4D split tensor, along the axis
-        # TODO: reinstate this test of uneven dimensions distribution
-        # after update to Allgatherv implementation (Issue  #273 depending on #233)
-        # data = ht.array(ht.random.randn(1, 4, 5, 1), split=1)
-        # result = ht.squeeze(data, axis=-1)
-        # self.assertIsInstance(result, ht.DNDarray)
-        # # TODO: the following works locally but not when distributed,
-        # #self.assertEqual(result.dtype, ht.float32)
-        # #self.assertEqual(result._DNDarray__array.dtype, torch.float32)
-        # self.assertEqual(result.shape, (1, 12, 5))
-        # self.assertEqual(result.lshape, (1, 12, 5))
-        # self.assertEqual(result.split, 1)
+        data = ht.array(ht.random.randn(1, 4, 5, 1), split=1)
+        result = ht.squeeze(data, axis=-1)
+        self.assertIsInstance(result, ht.DNDarray)
+        self.assertEqual(result.dtype, ht.float32)
+        self.assertEqual(result._DNDarray__array.dtype, torch.float32)
+        self.assertEqual(result.shape, (1, 4, 5))
+        self.assertEqual(result.split, 1)
 
         # 3D split tensor, across the axis
-        size = ht.MPI_WORLD.size * 2
+        size = ht.MPI_WORLD.size
         data = ht.triu(ht.ones((1, size, size), split=1, device=ht_device), k=1)
 
         result = ht.squeeze(data, axis=0)
         self.assertIsInstance(result, ht.DNDarray)
-        # TODO: the following works locally but not when distributed,
-        # self.assertEqual(result.dtype, ht.float32)
-        # self.assertEqual(result._DNDarray__array.dtype, torch.float32)
+        self.assertEqual(result.dtype, ht.float32)
+        self.assertEqual(result._DNDarray__array.dtype, torch.float32)
         self.assertEqual(result.shape, (size, size))
-        self.assertEqual(result.lshape, (size, size))
-        # self.assertEqual(result.split, None)
+        self.assertEqual(result.lshape, (1, size))
+        self.assertEqual(result.split, 0)
 
         # check exceptions
-        with self.assertRaises(ValueError):
-            data.squeeze(axis=(0, 1))
+        if data.comm.is_distributed():
+            with self.assertRaises(ValueError):
+                data.squeeze(axis=(0, 1))
         with self.assertRaises(TypeError):
             data.squeeze(axis=1.1)
         with self.assertRaises(TypeError):